JWT adicionado

+++ api-drugovich/app/Http/Controllers/Api/GroupController.php
@@ -6,13 +6,16 @@ use App\Http\Controllers\Controller;
 use App\Models\Client;
 use App\Models\Group;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class GroupController extends Controller
 {
     private $group;
+    private $loggedUser;
 
     public function __construct(Group $group)
     {
+        $this->loggedUser = Auth::user();
         $this->group = $group;
     }
     /**
@@ -33,7 +36,11 @@ class GroupController extends Controller
      */
     public function store(Request $request)
     {
-        return $this->group->create($request->all());
+        if($this->loggedUser->level == 2){
+            return $this->group->create($request->all());
+        }else{
+            return response()->json(['error' => 'You are not authorized to create groups'], 401);
+        }
     }
 
     /**
@@ -56,8 +63,12 @@ class GroupController extends Controller
      */
     public function update(Request $request, Group $group)
     {
-        $group->update($request->all());
-        return $group;
+        if($this->loggedUser->level == 2){
+            $group->update($request->all());
+            return response()->json($group, 200);
+        }else{
+            return response()->json(['error' => 'You are not authorized to update groups'], 401);
+        }
     }
 
     /**
@@ -68,16 +79,21 @@ class GroupController extends Controller
      */
     public function destroy(Group $group)
     {
-        if ($group != null) {
-            $clients = $group->clients;
-            foreach ($clients as $client) {
-                $client->group_id = null;
-                $client->save();
+        if($this->loggedUser->level == 2){
+            if ($group != null) {
+                $clients = $group->clients;
+                foreach ($clients as $client) {
+                    $client->group_id = null;
+                    $client->save();
+                }
+                $group->delete();
+                return response()->json(['message' => 'Group deleted']);
+            }else{
+                return response()->json(['message' => 'Group not found']);
             }
-            $group->delete();
-            return response()->json(['message' => 'Group deleted']);
         }else{
-            return response()->json(['message' => 'Group not found']);
+            return response()->json(['message' => 'You are not authorized to delete groups']);
         }
+        
     }
 }
